+++ src/main/java/chess/domain/board/BoardFactory.java
@@ -1,13 +1,14 @@
-package chess;
+package chess.domain.board;
 
-import chess.piece.Bishop;
-import chess.piece.Color;
-import chess.piece.King;
-import chess.piece.Knight;
-import chess.piece.Pawn;
-import chess.piece.Piece;
-import chess.piece.Queen;
-import chess.piece.Rook;
+import chess.domain.piece.King;
+import chess.domain.piece.Knight;
+import chess.domain.piece.Pawn;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Queen;
+import chess.domain.piece.Rook;
+import chess.domain.position.Position;
+import chess.domain.piece.Bishop;
+import chess.domain.piece.Color;
 import java.util.HashMap;
 import java.util.Map;
