+++ src/main/java/chess/domain/position/Path.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.domain.position;
 
 import java.util.ArrayList;
 import java.util.List;
