+++ src/main/java/chess/domain/position/Position.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.domain.position;
 
 import java.util.Objects;
