+++ src/test/java/chess/domain/board/BoardTest.java
@@ -1,11 +1,13 @@
-package chess;
+package chess.domain.board;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import chess.piece.Color;
-import chess.view.OutputView;
+import chess.domain.board.Board;
+import chess.domain.board.BoardFactory;
+import chess.domain.piece.Color;
+import chess.domain.position.Position;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
