+++ src/test/java/chess/domain/piece/KnightTest.java
@@ -1,11 +1,13 @@
-package chess.piece;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
-import chess.Path;
-import chess.Position;
+import chess.domain.piece.Color;
+import chess.domain.piece.Knight;
+import chess.domain.piece.Piece;
+import chess.domain.position.Path;
+import chess.domain.position.Position;
 import java.util.Optional;
 import org.assertj.core.api.InstanceOfAssertFactories;
 import org.junit.jupiter.api.Test;
