+++ src/test/java/chess/domain/piece/RookTest.java
@@ -1,23 +1,26 @@
-package chess.piece;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import chess.Path;
-import chess.Position;
+import chess.domain.piece.Color;
+import chess.domain.piece.King;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Rook;
+import chess.domain.position.Path;
+import chess.domain.position.Position;
 import java.util.Optional;
 import org.assertj.core.api.InstanceOfAssertFactories;
 import org.junit.jupiter.api.Test;
 
-class QueenTest {
+class RookTest {
 
     @Test
     void test_searchPathTo() {
-
-        Piece queen = new Queen(Color.WHITE);
+        Piece Rook = new Rook(Color.WHITE);
 
         Position initialPosition = new Position(5, 1);
-        Path path = queen.searchPathTo(initialPosition, new Position(5, 8), Optional.empty());
+        Path path = Rook.searchPathTo(initialPosition, new Position(5, 8), Optional.empty());
 
         assertThat(path)
                 .extracting("positions", InstanceOfAssertFactories.list(Position.class))
@@ -29,24 +32,10 @@ class QueenTest {
 
     @Test
     void test_searchPathTo2() {
-
-        Queen queen = new Queen(Color.WHITE);
-
-        Position initialPosition = new Position(5, 1);
-        Path path = queen.searchPathTo(initialPosition, new Position(8, 4), Optional.empty());
-
-        assertThat(path)
-                .extracting("positions", InstanceOfAssertFactories.list(Position.class))
-                .containsExactly(new Position(6, 2), new Position(7, 3));
-    }
-
-    @Test
-    void test_searchPathTo3() {
-
-        Queen queen = new Queen(Color.WHITE);
+        Rook Rook = new Rook(Color.WHITE);
 
         Position initialPosition = new Position(5, 5);
-        Path path = queen.searchPathTo(initialPosition, new Position(5, 1), Optional.empty());
+        Path path = Rook.searchPathTo(initialPosition, new Position(5, 1), Optional.empty());
 
         assertThat(path)
                 .extracting("positions", InstanceOfAssertFactories.list(Position.class))
@@ -57,14 +46,13 @@ class QueenTest {
     }
 
     @Test
-    void test_searchPathTo4() {
-
-        Queen queen = new Queen(Color.WHITE);
+    void test_searchPathTo3() {
+        Rook Rook = new Rook(Color.WHITE);
 
         Position initialPosition = new Position(5, 5);
 
         assertThatThrownBy(
-                () -> queen.searchPathTo(initialPosition,
+                () -> Rook.searchPathTo(initialPosition,
                         new Position(5, 1),
                         Optional.of(new King(Color.WHITE))))
                 .isInstanceOf(IllegalStateException.class);
