+++ src/test/java/chess/domain/position/PositionTest.java
@@ -1,7 +1,9 @@
-package chess;
+package chess.domain.position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import chess.domain.position.Movement;
+import chess.domain.position.Position;
 import org.junit.jupiter.api.Test;
 
 class PositionTest {
